fix(port_riscv): match the size of `TaskState::sp` to XLEN

+++ src/constance_port_riscv/src/threading/imp.rs
@@ -298,7 +298,7 @@ unsafe impl Sync for State {}
 #[derive(Debug)]
 #[repr(C)]
 pub struct TaskState {
-    sp: UnsafeCell<u32>,
+    sp: UnsafeCell<usize>,
 }
 
 unsafe impl Sync for TaskState {}
